requirements.txt

# NOTE: Built on python 3.12.3

fastapi[app]
pydantic == 2.12.3
psycopg2-binary==2.9.11
SQLAlchemy==2.0.44
uvicorn==0.38.0
alembic==1.17.0
bcrypt==4.0.1
langgraph==1.0.1
langchain==1.0.2
langchain-openai==1.0.1
openai==2.6.1
asyncpg==0.30.0
colorama==0.4.6
passlib==1.7.4
dotenv==0.9.9
python-multipart==0.0.20
PyJWT==2.10.1